# PERF: Use backend to load `Note`