--- a/AnkiDroid/src/main/java/com/ichi2/libanki/Note.kt
+++ b/AnkiDroid/src/main/java/com/ichi2/libanki/Note.kt
@@ -47,7 +47,6 @@ class Note : Cloneable {
         private set
     lateinit var fields: MutableList<String>
         private set
-    private var mFlags = 0
     private var mData: String? = null
     private var mFMap: Map<String, Pair<Int, JSONObject>>? = null
     var usn = 0
@@ -69,7 +68,6 @@ class Note : Cloneable {
         mid = notetype.getLong("id")
         tags = mutableListOf()
         fields = emptyStringMutableList(notetype.getJSONArray("flds").length())
-        mFlags = 0
         mData = ""
         mFMap = Notetypes.fieldMap(this.notetype)
     }
@@ -89,7 +87,6 @@ class Note : Cloneable {
                 usn = cursor.getInt(3)
                 tags = col.tags.split(cursor.getString(4))
                 fields = Utils.splitFields(cursor.getString(5))
-                mFlags = cursor.getInt(6)
                 mData = cursor.getString(7)
                 notetype = col.notetypes.get(mid)!!
                 mFMap = Notetypes.fieldMap(notetype)