--- a/AnkiDroid/src/main/java/com/ichi2/utils/StringUtil.kt
+++ b/AnkiDroid/src/main/java/com/ichi2/utils/StringUtil.kt
@@ -42,4 +42,6 @@ fun String.lastIndexOfOrNull(c: Char): Int? =
         else -> index
     }
 
+fun emptyStringMutableList(size: Int): MutableList<String> = MutableList(size) { "" }
+
 fun emptyStringArray(size: Int): Array<String> = Array(size) { "" }