## Purpose / Description
Match Anki Desktop's implementation, and get a 3x speedup (1m30 -> 30s to load 1MM notes)

❗️WARN: while unit testing, I had a failure of:

* https://github.com/ankidroid/Anki-Android/issues/14796

I have annotated the class which I believe is the problem, but this is not guaranteed to fix the issue. I have performed many retests and have been unable to reproduce the test failure

## Approach
* try to use the backend, find there are type mismatches
* fix the mismatches
* use the backend
* performance test


## How Has This Been Tested?

1m 27 -> 28s to load 1 million notes (~87 -> 27 micros)

```kotlin
    @Test
    fun notePerformance() {
        val note = addNoteUsingBasicModel("hello", "world")

        // 28s 271ms    - backend + no .toMutableList()
        // 28s 461ms    - backend + .toMutableList()
        // 1m 27s       - Cursor
        for (i in 0..1_000_000) {
            col.getNote(note.id)
        }
    }
```

## Learning (optional, can help others)

https://github.com/ankitects/anki/blob/db93939ded947d74bb25bd8552a2b2356a096509/pylib/anki/notes.py#L54-L67

Backend Codegen is problematic: it exposes a `java.util.List`, but this is actually a `com.google.protobuf.Internal.ProtobufList`

> Extends List to add the capability to make the list immutable and inspect if it is modifiable.

https://protobuf.dev/reference/java/api-docs/com/google/protobuf/Internal.ProtobufList


## Checklist
- [x] You have a descriptive commit message with a short title (first line, max 50 chars).
- [x] You have commented your code, particularly in hard-to-understand areas
- [x] You have performed a self-review of your own code
- [ ] UI changes: include screenshots of all affected screens (in particular showing any new or changed strings)
- [ ] UI Changes: You have tested your change using the [Google Accessibility Scanner](https://play.google.com/store/apps/details?id=com.google.android.apps.accessibility.auditor)