@mikehardy 2023-12-21
This one gave me a fit locally. 2 macOS machines ran tests with this PR rebased to main no problem but windows consistent failures. Stopped gradle daemons, gradle clean with `--rerun-tasks` then it was clean. Poisoned cache or something. Hate that "I'm missing something..." feeling that gives me.