@mikehardy 2023-12-21
Nope, it's this, sometimes, I cannot get it to reproduce but I did just see it on a macOS machine so I've seen it cross-platform now. CI has zero instances, and locally I'm green again, I dunno

```
ReviewerTest > testLrnQueueAfterUndo FAILED
    java.lang.NoSuchMethodError: 'java.lang.String[] com.ichi2.libanki.Note.getFields()'
        at com.ichi2.anki.RobolectricTest.equalFirstField(RobolectricTest.kt:427)
        at com.ichi2.anki.ReviewerTest$testLrnQueueAfterUndo$1.invokeSuspend(ReviewerTest.kt:166)
        at com.ichi2.anki.ReviewerTest$testLrnQueueAfterUndo$1.invoke(ReviewerTest.kt)
        at com.ichi2.anki.ReviewerTest$testLrnQueueAfterUndo$1.invoke(ReviewerTest.kt)
        at com.ichi2.testutils.TestClass$runTest$1.invokeSuspend(TestClass.kt:154)
        at com.ichi2.testutils.TestClass$runTest$1.invoke(TestClass.kt)
        at com.ichi2.testutils.TestClass$runTest$1.invoke(TestClass.kt)
        at kotlinx.coroutines.test.TestBuildersKt__TestBuildersKt$runTest$2$1$1.invokeSuspend(TestBuilders.kt:314)
        at kotlin.coroutines.jvm.internal.BaseContinuationImpl.resumeWith(ContinuationImpl.kt:33)
        at kotlinx.coroutines.DispatchedTask.run(DispatchedTask.kt:108)
        at kotlinx.coroutines.test.TestDispatcher.processEvent$kotlinx_coroutines_test(TestDispatcher.kt:28)
        at kotlinx.coroutines.test.TestCoroutineScheduler.tryRunNextTaskUnless$kotlinx_coroutines_test(TestCoroutineScheduler.kt:103)
        at kotlinx.coroutines.test.TestBuildersKt__TestBuildersKt$runTest$2$1$workRunner$1.invokeSuspend(TestBuilders.kt:320)
        at kotlin.coroutines.jvm.internal.BaseContinuationImpl.resumeWith(ContinuationImpl.kt:33)
        at kotlinx.coroutines.DispatchedTask.run(DispatchedTask.kt:108)
        at kotlinx.coroutines.EventLoopImplBase.processNextEvent(EventLoop.common.kt:280)
        at kotlinx.coroutines.BlockingCoroutine.joinBlocking(Builders.kt:85)
        at kotlinx.coroutines.BuildersKt__BuildersKt.runBlocking(Builders.kt:59)
        at kotlinx.coroutines.BuildersKt.runBlocking(Unknown Source)
        at kotlinx.coroutines.BuildersKt__BuildersKt.runBlocking$default(Builders.kt:38)
        at kotlinx.coroutines.BuildersKt.runBlocking$default(Unknown Source)
        at kotlinx.coroutines.test.TestBuildersJvmKt.createTestResult(TestBuildersJvm.kt:13)
        at kotlinx.coroutines.test.TestBuildersKt__TestBuildersKt.runTest-8Mi8wO0(TestBuilders.kt:308)
        at kotlinx.coroutines.test.TestBuildersKt.runTest-8Mi8wO0(Unknown Source)
        at kotlinx.coroutines.test.TestBuildersKt__TestBuildersKt.runTest-8Mi8wO0(TestBuilders.kt:166)
        at kotlinx.coroutines.test.TestBuildersKt.runTest-8Mi8wO0(Unknown Source)
        at com.ichi2.testutils.TestClass.runTest(TestClass.kt:152)
        at com.ichi2.testutils.TestClass.runTest$default(TestClass.kt:146)
        at com.ichi2.anki.ReviewerTest.testLrnQueueAfterUndo(ReviewerTest.kt:150)

```